requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.8.2
    # via flask
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via flask
colorama==0.4.6
    # via
    #   click
    #   tqdm
dash==2.18.2
    # via
    #   -r requirements.in
    #   dash-auth
dash-auth==2.3.0
    # via -r requirements.in
dash-core-components==2.0.0
    # via
    #   -r requirements.in
    #   dash
dash-html-components==2.0.0
    # via
    #   -r requirements.in
    #   dash
dash-table==5.0.0
    # via
    #   -r requirements.in
    #   dash
flask==3.0.3
    # via
    #   dash
    #   dash-auth
idna==3.10
    # via requests
importlib-metadata==8.5.0
    # via dash
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
nest-asyncio==1.6.0
    # via dash
numpy==2.1.3
    # via
    #   -r requirements.in
    #   pandas
packaging==24.1
    # via plotly
pandas==2.2.3
    # via -r requirements.in
plotly==5.24.1
    # via dash
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via pandas
requests==2.32.3
    # via dash
retrying==1.3.4
    # via dash
six==1.16.0
    # via
    #   python-dateutil
    #   retrying
tenacity==9.0.0
    # via plotly
tqdm==4.66.6
    # via -r requirements.in
typing-extensions==4.12.2
    # via dash
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
werkzeug==3.0.6
    # via
    #   dash
    #   dash-auth
    #   flask
zipp==3.20.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
